Refactor out test utils
